+++ fuel-crypto/src/secp256/public.rs
@@ -1,6 +1,6 @@
 use crate::{
     hasher::Hasher,
-    secp::SecretKey,
+    secp256::SecretKey,
     Error,
 };
 use core::{
